💡 refactor(GamePage): extract map logic to MapContainer
Extracted map-related logic to a new `MapContainer` component for better modularity and code reuse. Updated `GamePage` to utilize the new component, simplifying its structure and improving maintainability.
